fix for test-reduced-permissions and add disableClear for login commands

+++ tests/integration/support/login-commands.js
@@ -7,9 +7,11 @@ const PASSWORD = Cypress.env('OIDC_PASS');
 
 Cypress.Commands.add('loginAndSelectCluster', function(params) {
   cy.handleExceptions();
-  sessionStorage.clear();
-  cy.clearCookies();
-  cy.clearLocalStorage();
+  if (!params?.disableClear) {
+    sessionStorage.clear();
+    cy.clearCookies();
+    cy.clearLocalStorage();
+  }
 
   const defaults = {
     fileName: 'kubeconfig.yaml',
